requirements.txt
#install with "pip3 install -r requirements.txt"

discord.py
youtube_dl
colour
psutil
bs4
requests
Pillow
pytz
sqlalchemy
coloredlogs